refactor: Extract gene tools into separate module
- Create src/tools/gene_tools.py with 3 gene query functions
- Move gene tools from server.py to dedicated module:
  - get_gene_by_entrez_id()
  - get_gene_by_symbol()
  - get_gene_by_position()
- Add register_tools() function for clean registration
- Update server.py to import and register gene tools

Benefits:
- Clear separation of gene-related functionality
- Easy to extend with new gene tools
- Reduced server.py from ~775 to ~650 lines
- Better code organization and maintainability

All existing functionality preserved - backward compatible.
