+++ src/tools/__init__.py
@@ -1,0 +1,7 @@
+"""
+Tool modules for MARRVEL-MCP.
+
+This package contains all the MCP tool implementations organized by category.
+"""
+
+__all__ = ["gene_tools"]
